requirements.txt
bibtexparser==1.2.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
feedparser==5.2.1
Flask==1.1.2
future==0.18.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pathlib==1.0.1
pyparsing==2.4.7
python-dotenv==0.13.0
pytz==2020.1
Pyzotero==1.4.16
requests==2.23.0
urllib3==1.25.9
Werkzeug==1.0.1
